SAS-48 Перенести логику аутентификации из контроллера в UserService
