+++ core/src/main/java/ru/gb/stalser/core/services/interfaces/UserService.java
@@ -1,7 +1,10 @@
 package ru.gb.stalser.core.services.interfaces;
 
 
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.userdetails.UserDetailsService;
+import ru.gb.stalser.api.dto.auth.AuthRequest;
+import ru.gb.stalser.api.dto.auth.AuthResponse;
 import ru.gb.stalser.core.entity.User;
 
 import java.util.List;
@@ -19,5 +22,8 @@ public interface UserService extends UserDetailsService {
     User save(User user);
 
     void updateUser(User user);
+
     Boolean existsByEmail(String email);
+
+    ResponseEntity<AuthResponse> authenticate(AuthRequest authRequest);
 }
